backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
psycopg2-binary
asyncpg
pgvector
slack-sdk
google-auth
google-auth-oauthlib
google-auth-httplib2
langchain
langchain-openai
langchain-text-splitters
openai
python-multipart
python-jose[cryptography]
passlib[bcrypt]
alembic
apscheduler
pydantic-settings
pytest
httpx
pytest-asyncio
